feat: refactor to ruby 3.1

+++ db/migrate/20210104060422_create_certificates.rb
@@ -10,7 +10,7 @@ class CreateCertificates < ActiveRecord::Migration[6.1]
       t.text :alternate_names
       t.string :requester
       t.boolean :validated, null: false, default: false, index: true
-      t.integer :serial, null: false, index: { unique: true }
+      t.bigint :serial, null: false, index: { unique: true }
 
       t.timestamps
     end
